POC: Search improvements (#98)
I'm experimenting with a few things for the search, trying to make sense
of Algolia's docs and how to achieve our aims with it. What I'm trying:

- Experimenting with reformatting the search result URL with
`transformItems`, to remove the domain in the item URL. The goal here is
for the search to work in local dev envs
- Set an explicit class on the main content following [the advice
here](https://docsearch.algolia.com/docs/required-configuration#use-the-right-classes-as-recordprops).
I plan to set this in the crawler configuration as the `content`
selector. This will hopefully ensure we capture the breadcrumbs and the
H1s in the index, as previously I was using the `.content` selector for
this.
- Add the organisation name (ie: ROM, Dry, Hanami) to the top of the
sidebar on the Guides page. Set an explicit `lvl1` class on this, so we
can set it as the lvl1 property in the index. I want to see what result
this gives us when we reindex.
- Add an explicit `lvl2` class to the doc/guide version, so that we can
set this as the lvl2 property in the index. I note that we already have
the version in a meta tag, which I understand we can use for
filtering/faceting, but I also want to try showing the version number
for a guide/doc in the search results dialog.

eg: for a Guide page like this:
https://hanakai.org/guides/hanami/v2.2/database/configuration

If I search for "DATABASE_URL", it would be cool to see in the search
results that text in the context of that page and some indication that
the page lives under Guides > Hanami > 2.2 > Database