Implement header for desktop/mobile in new design (#101)
This PR adds the implementation of the site header across desktop and
mobile. Notable things:

- **Mobile animation**
On mobile the header animates down nicely, but only in Chrome 🥲 I’m
trying to adhere to standard CSS things wherever possible, and so it
uses
[`interpolate-size`](https://piccalil.li/blog/the-interpolate-size-property-is-a-great-example-of-progressive-enhancement/)
to transition between `height: 0` and `height: auto`. As browser support
for that feature is released, it’ll be "upgraded" to the fanciness.
- **Tailwind compilation**
I’ve changed the Tailwind compilation to be through
[`postcss`](https://postcss.org) instead of our previous method of
pre-compiling Tailwind and then `@import`-ing the built files. I wanted
this so we can properly use Tailwind’s `@variant` directives for custom
media queries (which otherwise would’ve involved spreading out the
definitions across the codebase).
  
To do this I had to write my own postcss plugin 🙃 but the upside is that
it all works perfectly now.
- **`toggleClass` view function**
I’ve added a new `toggle-class` view function that allows us to
declaratively attach behaviour to an element to toggle its own classes
(or the classes of other matching elements).
- **Dynamic nav**
The navigation items are resolved from the main `View` class now, and
use a simple `Data` object to define their structure.

Follow-up work:

- Change theme based on project being viewed
- Fix up Docsearch theme
- Add in proper logo SVGs