Handle trailing slashes with rack-rewrite (#123)
Right now if we visit a guide page like
/guides/hanami/v2.2/getting-started with a trailing slash at the end
then Puma will 404. This isn't currently a problem in production due to
the CDN doing the same thing, but if we ever deploy this elsewhere it
might become a problem and it's certainly unnecessarily annoying in
development mode.

Extracted from #120 and #122

Co-authored-by: Josephine Hall <josephinehall@gmail.com>