Drop docs, centre on guides, with different versioning schemes (#197)
After some discussion with @makenosound, we feel like the distinction
between "guides" and "docs" is not a helpful one. It will be confusing
for our experience, provide a worse overall experience, and end up with
us having to maintain a bunch of redundant documentation if we wanted
both sides to be fleshed out.

So instead of this, we're dropping docs and centring everything around
guides.

In this change, we:

- Move the Dry per-gem docs to become guides
- Delete all other docs
- Remove all the code for loading/displaying docs
- Support different versioning schemes for guides: they can be
_org-versioned_ (e.g. for Hanami and Rom), _self-versioned_ (e.g. for
the Dry gem docs) or _unversioned_ (e.g. for the cross-cutting Dry
guides). Each org is either (a) org-versioned (Hanami and Rom), or (b) a
mix of self-versioned and un-versioned (Dry).

The result is intuitive: all the guides are available at /guides, and
you can just click through to read what you want.

The Hanami and rom guides continue to work like before, with an org
version selector at the top of the LHS nav.

The Dry guides are a little different. The LHS nav gets the latest
version for all the self-versioned guides, and mixes in the unversioned
guides at their respective positions. There is no org level version
selector at the top of the LHS nav. Instead, @makenosound is going to
find a different UI solution for providing a version selector when
reading a self-versioned guide, by e.g. putting a version selector next
to that guide's heading in the LHS nav. The ultimate approach for this
is TBD and @makenosound will figure it out when he gets to it :)

In a future PR, we should change the /guides URL to /learn (and update
the top nav item to match). That feels like a nicer name for this area.
We don't need to expose our underlying database model to the users like
we have so far. Moving away from this is the whole point of this PR.