Push-based ContentFileMiddleware (#195)
Following up on #187, this is my idea of simplifying (?)
`ContentFileMiddleware`.

## Rationale

Current static files handling operates in kind of two modes:

1. "Pull mode", where upon receiving a URL, the Rack middleware maps it
to a path on disk and serves. This is used in dev.
2. "Push mode", where filesystem is scanned, URLs generated and content
copied to reflect the URLs. This is used in deployment.

This leads to errors that are quite hard to catch without manual testing
for second approach, because it's only used when actually building a
site. I thought it would be better to unify the modes and the "push
mode" was the only option here, as we don't have Rack middleware during
the deployment.

## Solution

The `ContentFileMiddleware` eagerly scans the filesystem, according to
provided rules, in search of asset files. Then it builds a map of `URL
=> path`, which is stored in memory. This happens only on first request
of a file with supported extension, to avoid scanning on every change in
dev.

When a request is made, the middleware just scans the map for known URLs
and serves the file from `path` if something is found.

Similarly, during the build, the middleware's map-building process is
reused to find all the files to copy. The map is inverted to become path
=> URL map, and then the files are copied to `build` directory to be
served under the matching URL.

This allows much more flexible mapping between the path and the URL, in
case we want to bring back the assets structure from current Hanami
guides page, where assets are stored alongside the posts markdown, not
in a separate directory.