Order Dry guides alphabetically, show deprecated (#232)
Alphabetical ordering makes it easiest to find other guides, I think.
"Getting started" remains at the top of the list, of course. In this
area we'll add a more themed/categorised set of links to all the guides.

Put deprecated guides at the end of the list, and mark them as
"deprecated".