Fix majority of 404s (#234)
Hello! This PR solves _most_ of the 404s raised during the site
crawling, however I have a few notes on what I've done:

- Updated all stray links like `/gems/dry-transaction` to
`//doc/dry-transaction`, however I think we may need to change the path
format for these across the board to `//guide`, with appropriate
handling for the `dry` family? These seem to make up the remaining 404s
that come up during the crawling. As a stop-gap, I could update them to
`//doc/dry/dry-transaction` if you like?
- Updated a number `//guide` links to go to correct places within the
same version
- In _old_ blog posts that reference Lotus etc, I've just removed a
number of links as we don't have those original versions of the docs to
link to. LMK if you'd prefer a different approach here, I'm not sure
where else these could link to.